# fix(tasks): Reduce allocations in scheduler