@stuartcarnie 2020-02-07
Wrote some benchmarks to demonstrate the resource improvements

```
↳ benchcmp before.txt after.txt
benchmark                                                             old ns/op     new ns/op     delta
BenchmarkTreeScheduler/10_tasks_per_worker/100µs_worker_delay-16      2666342       2633051       -1.25%
BenchmarkTreeScheduler/10_tasks_per_worker/500µs_worker_delay-16      7143547       7116340       -0.38%
BenchmarkTreeScheduler/10_tasks_per_worker/1ms_worker_delay-16        12532812      12617855      +0.68%
BenchmarkTreeScheduler/50_tasks_per_worker/100µs_worker_delay-16      8829934       8194148       -7.20%
BenchmarkTreeScheduler/50_tasks_per_worker/500µs_worker_delay-16      33721751      33446650      -0.82%
BenchmarkTreeScheduler/50_tasks_per_worker/1ms_worker_delay-16        63691206      63841880      +0.24%
BenchmarkTreeScheduler/100_tasks_per_worker/100µs_worker_delay-16     17518004      15466442      -11.71%
BenchmarkTreeScheduler/100_tasks_per_worker/500µs_worker_delay-16     68403143      66198825      -3.22%
BenchmarkTreeScheduler/100_tasks_per_worker/1ms_worker_delay-16       129496178     126765651     -2.11%

benchmark                                                             old allocs     new allocs     delta
BenchmarkTreeScheduler/10_tasks_per_worker/100µs_worker_delay-16      4769           2149           -54.94%
BenchmarkTreeScheduler/10_tasks_per_worker/500µs_worker_delay-16      29106          14641          -49.70%
BenchmarkTreeScheduler/10_tasks_per_worker/1ms_worker_delay-16        65405          30991          -52.62%
BenchmarkTreeScheduler/50_tasks_per_worker/100µs_worker_delay-16      26284          7254           -72.40%
BenchmarkTreeScheduler/50_tasks_per_worker/500µs_worker_delay-16      116106         33849          -70.85%
BenchmarkTreeScheduler/50_tasks_per_worker/1ms_worker_delay-16        223882         68148          -69.56%
BenchmarkTreeScheduler/100_tasks_per_worker/100µs_worker_delay-16     47370          9010           -80.98%
BenchmarkTreeScheduler/100_tasks_per_worker/500µs_worker_delay-16     208674         42368          -79.70%
BenchmarkTreeScheduler/100_tasks_per_worker/1ms_worker_delay-16       406686         83153          -79.55%

benchmark                                                             old bytes     new bytes     delta
BenchmarkTreeScheduler/10_tasks_per_worker/100µs_worker_delay-16      750866        106131        -85.87%
BenchmarkTreeScheduler/10_tasks_per_worker/500µs_worker_delay-16      4538768       707587        -84.41%
BenchmarkTreeScheduler/10_tasks_per_worker/1ms_worker_delay-16        10140374      1491921       -85.29%
BenchmarkTreeScheduler/50_tasks_per_worker/100µs_worker_delay-16      6729135       361175        -94.63%
BenchmarkTreeScheduler/50_tasks_per_worker/500µs_worker_delay-16      29625844      1647872       -94.44%
BenchmarkTreeScheduler/50_tasks_per_worker/1ms_worker_delay-16        56760879      3294103       -94.20%
BenchmarkTreeScheduler/100_tasks_per_worker/100µs_worker_delay-16     13277604      458190        -96.55%
BenchmarkTreeScheduler/100_tasks_per_worker/500µs_worker_delay-16     58079158      2079924       -96.42%
BenchmarkTreeScheduler/100_tasks_per_worker/1ms_worker_delay-16       113214325     4040935       -96.43%
```